// src/components/main-product/main-product.types.d.ts
export interface Settings {
    cashback_text: string;
    coupon_code: string;
    coupon_label: string;
    shipping_text: string;
    description_title: string;
    bundle_use_default: any;
    bundle_product: string;
    bundle_discount: any;
    bundle_text: string;
    benefits_svg_1: string;
    benefits_label_1: string;
    benefits_svg_2: string;
    benefits_label_2: string;
    benefits_svg_3: string;
    benefits_label_3: string;
    fixed_collection: any;
    fixed_collection_collection: string;
    filter_by_crystal: any;
}